# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediFlow — Multi-tenant AI Healthcare Appointment Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  admin: 'admin',
  doctor: 'doctor',
  patient: 'patient',
} as const;

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  tenantId: number;
  role: RegisterInputRole;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  doctor: 'doctor',
  patient: 'patient',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  tenantId: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Tenant {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  createdAt: string;
}

export interface TenantInput {
  name: string;
  slug: string;
  address?: string;
  phone?: string;
  email?: string;
  logoUrl?: string;
}

export interface TenantUpdate {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  logoUrl?: string;
}

export interface Doctor {
  id: number;
  tenantId: number;
  name: string;
  specialty: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  availableDays?: string;
  startTime?: string;
  endTime?: string;
  slotDurationMinutes?: number;
  createdAt: string;
}

export interface DoctorInput {
  tenantId: number;
  name: string;
  specialty: string;
  bio?: string;
  licenseNumber?: string;
  phone?: string;
  email?: string;
  availableDays?: string;
  startTime?: string;
  endTime?: string;
  slotDurationMinutes?: number;
}

export interface DoctorUpdate {
  name?: string;
  specialty?: string;
  bio?: string;
  licenseNumber?: string;
  phone?: string;
  email?: string;
  availableDays?: string;
  startTime?: string;
  endTime?: string;
  slotDurationMinutes?: number;
}

export interface TimeSlot {
  date: string;
  startTime: string;
  endTime: string;
  available: boolean;
}

export interface Patient {
  id: number;
  tenantId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PatientInput {
  tenantId: number;
  name: string;
  email?: string;
  phone?: string;
  dateOfBirth?: string;
  bloodType?: string;
  address?: string;
  notes?: string;
}

export interface PatientUpdate {
  name?: string;
  email?: string;
  phone?: string;
  dateOfBirth?: string;
  bloodType?: string;
  address?: string;
  notes?: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  scheduled: 'scheduled',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Appointment {
  id: number;
  tenantId: number;
  doctorId: number;
  patientId: number;
  appointmentDate: string;
  startTime: string;
  endTime: string;
  status: AppointmentStatus;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  specialty?: string | null;
  createdAt: string;
}

export interface AppointmentInput {
  tenantId: number;
  doctorId: number;
  patientId: number;
  appointmentDate: string;
  startTime: string;
  endTime: string;
  reason?: string;
  notes?: string;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  scheduled: 'scheduled',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface AppointmentUpdate {
  appointmentDate?: string;
  startTime?: string;
  endTime?: string;
  reason?: string;
  notes?: string;
  status?: AppointmentUpdateStatus;
}

export interface CancelInput {
  cancelReason?: string;
}

export interface CompleteInput {
  notes?: string;
}

export interface SpecialtyCount {
  specialty: string;
  count: number;
}

export interface DashboardStats {
  totalAppointments: number;
  scheduledAppointments: number;
  completedAppointments: number;
  cancelledAppointments: number;
  totalDoctors: number;
  totalPatients: number;
  todayAppointments: number;
  appointmentsBySpecialty: SpecialtyCount[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  appointment_booked: 'appointment_booked',
  appointment_cancelled: 'appointment_cancelled',
  appointment_completed: 'appointment_completed',
  patient_registered: 'patient_registered',
  doctor_added: 'doctor_added',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  doctorName?: string | null;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatInput {
  message: string;
  tenantId: number;
  conversationHistory?: ChatMessage[];
}

export type AiReplyActionData = { [key: string]: unknown };

export interface AiReply {
  message: string;
  suggestions: string[];
  /** @nullable */
  action?: string | null;
  actionData?: AiReplyActionData;
}

export type ListDoctorsParams = {
tenantId?: number;
specialty?: string;
};

export type ListPatientsParams = {
tenantId?: number;
search?: string;
};

export type ListAppointmentsParams = {
tenantId?: number;
doctorId?: number;
patientId?: number;
status?: ListAppointmentsStatus;
date?: string;
};

export type ListAppointmentsStatus = typeof ListAppointmentsStatus[keyof typeof ListAppointmentsStatus];


export const ListAppointmentsStatus = {
  scheduled: 'scheduled',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export type GetDashboardStatsParams = {
tenantId?: number;
};

export type GetDashboardActivityParams = {
tenantId?: number;
limit?: number;
};