# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediFlow — Multi-tenant AI Healthcare Appointment Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, cancelled, completed]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Booked appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInput"
      responses:
        "200":
          description: Cancelled appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      summary: Mark appointment as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteInput"
      responses:
        "200":
          description: Completed appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI appointment assistant chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReply"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, tenantId, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        tenantId:
          type: integer
        role:
          type: string
          enum: [admin, doctor, patient]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, tenantId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, doctor, patient]
        tenantId:
          type: integer
        createdAt:
          type: string
          format: date-time

    # Tenant
...
    Patient:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PatientInput:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        bloodType:
          type: string
        address:
          type: string
        notes:
          type: string
...
    Appointment:
      type: object
      required: [id, tenantId, doctorId, patientId, appointmentDate, startTime, endTime, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        doctorId:
          type: integer
        patientId:
          type: integer
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, cancelled, completed]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [tenantId, doctorId, patientId, appointmentDate, startTime, endTime]
      properties:
        tenantId:
          type: integer
        doctorId:
          type: integer
        patientId:
          type: integer
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: string
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [scheduled, cancelled, completed]
    CancelInput:
      type: object
      properties:
        cancelReason:
          type: string
    CompleteInput:
      type: object
      properties:
        notes:
          type: string

    # Dashboard
...
    AiChatInput:
      type: object
      required: [message, tenantId]
      properties:
        message:
          type: string
        tenantId:
          type: integer
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    AiReply:
      type: object
      required: [message, suggestions]
      properties:
        message:
          type: string
        suggestions:
          type: array
          items:
            type: string
        action:
          type: ["string", "null"]
        actionData:
          type: object