# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediFlow — Multi-tenant AI Healthcare Appointment Platform
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  appointmentDate?: Date;
  startTime?: string;
  endTime?: string;
  reason?: string;
  notes?: string;
  status?: AppointmentUpdateStatus;
}