# lib/api-zod/src/generated/types/cancelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediFlow — Multi-tenant AI Healthcare Appointment Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CancelInput {
  cancelReason?: string;
}